Add countdown state to protocol and server

+++ shared/src/resources/mod.rs
@@ -1,7 +1,7 @@
 /// Stores configuration on the map of a single game
 pub struct MapConfig {
     /// The number of tiles in width the map is
-    pub size_width: i32,
+    pub size_width: u16,
     /// The number of tiles in height the map is
-    pub size_height: i32,
+    pub size_height: u16,
 }
